+++ breve_drf_template/apps/employee/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
-# from breve_drf_template.apps.Employee.models import EmployeeModel
+# from breve_drf_template.apps.employee.models import EmployeeModel
 
 
 class UserConfig(AppConfig): 
@@ -8,5 +8,5 @@ class UserConfig(AppConfig):
 
 
 class EmployeeConfig(AppConfig):
-    name = 'breve_drf_template.apps.Employee'
+    name = 'breve_drf_template.apps.employee'
     verbose_name = 'Empleados'
